add setup.py, change directory structure
