+++ ironmarch/analysis/generate_embeddings.py
