+++ ironmarch/datasets/ironmarch.py
@@ -1,14 +1,6 @@
-from typing import Callable, List
-import torch
-import pandas as pd
-from torch import tensor
-from torch.utils import data
-from utils import chunk_list
-import numpy as np
-import csv
-import html2text
-
 from torch.utils.data import Dataset
+import pandas as pd
+import html2text
 import os
 
 class IronMarch(Dataset):
@@ -42,49 +34,6 @@ class IronMarch(Dataset):
         posts_csv['data'] = posts_csv['data'].map(lambda data: html2text.html2text(str(data)))
 
         self.data = dm_csv.to_dict(orient = 'records') + posts_csv.to_dict(orient = 'records')
-
-# Assumes that the dataset is using the {'id', 'data', 'label'} format
-class BertPreprocess():
-    def __init__(
-        self,
-        dataset: Dataset,
-        bert: torch.nn.Module,
-        tokenizer: Callable,
-        device = 'cpu',
-        batch_size = 1024,
-        **kwargs
-    ):
-        self.bert = bert.to(device).eval()
-        self.tokenizer = tokenizer
-        self.device = device
-
-        self.dataset = dataset
-        self.batch_size = batch_size
-        self.data = self._preprocess(self.dataset)
-
-    def __getitem__(self, index):
-        return self.data[index]
-
-    def __len__(self):
-        return len(self.data)
-
-    def _preprocess(self, dataset: Dataset) -> List[dict]:
-        unprocessed_data = []
-        for item in dataset:
-            unprocessed_data.append(item['data'])
-        for chunk in chunk_list(unprocessed_data, self.batch_size):
-            tokens = self._tokenize(chunk)
-            bert_embeddings = self.bert(**tokens)
-        
-    def _tokenize(self, strings_to_tokenize):
-        tokens = self.tokenizer(
-            strings_to_tokenize,
-            padding = 'max_length',
-            max_length = self.config['max_length'],
-            return_tensors = 'pt',
-            verbose = False
-        )
-        return {key : value[0][len(value[0]) - self.config['max_length']:] for key, value in tokens.items()}
 
 if __name__ == '__main__':
     dataset = IronMarch('data/iron-march')
