+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+if __name__ == '__main__':
+    setup(
+        name = 'ironmarch',
+        version = '0.0.1',
+        packages = find_packages()
+    )
